ingest/requirements.txt
aiokafka~=0.12.0
aiohttp~=3.12.14
python-dotenv~=1.1.0
prometheus_client~=0.22.1
requests~=2.32.3